Added functionality to delete Categories. Made a function that deletes a rule given an ID, used both by deleteRule.php and deleteCategory.php. Added href='#' to links to make them focusable.

+++ includes/deleteQuestion.php
@@ -31,6 +31,18 @@ mysqli_stmt_bind_param($stmtDeleteAnswers, "i", $questionID);
 mysqli_stmt_execute($stmtDeleteAnswers);
 mysqli_stmt_close($stmtDeleteAnswers);
 
+$sqlDeleteGrades = "DELETE FROM gradeperquestion WHERE questionID = ?;";
+$stmtDeleteGrades = mysqli_stmt_init($conn);
+
+if (!mysqli_stmt_prepare($stmtDeleteGrades, $sqlDeleteGrades)) {
+    echo ("error=stmtFailed");
+    exit();
+}
+
+mysqli_stmt_bind_param($stmtDeleteGrades, "i", $questionID);
+mysqli_stmt_execute($stmtDeleteGrades);
+mysqli_stmt_close($stmtDeleteGrades);
+
 $sqlDeleteQuestion = "DELETE FROM questions WHERE questionID = ?;";
 $stmtDeleteQuestion = mysqli_stmt_init($conn);
